requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1
pandas==2.0.3
requests==2.31.0
pydantic==2.5.0
# Local LLM packages
transformers==4.36.0
torch==2.1.0
accelerate==0.24.0
sentence-transformers==2.2.2
# NER and NLP packages
spacy==3.7.2
# Vector store
faiss-cpu==1.7.4
# OpenAI package (optional, for fallback)
openai==1.3.0
